contrib/google-cloud/store-api/requirements.txt
quart
hypercorn
google-cloud-firestore
